% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationMatrix.R
\name{correlationMatrix}
\alias{correlationMatrix}
\title{explore the correlations between the features}
\usage{
correlationMatrix(subgroup.data, method = "pearson")
}
\arguments{
\item{method}{How to calculate the correlation c("pearson", "spearman")}

\item{subgroup.dataset}{a list of data, covariates, outcomes and the trt, see loadDataset (list)}
}
\value{
a list or R and P values (list)
}
\description{
This function allows you to get the correlation between features
with either pearson or spearson method
using the rcorr function, which returns correlations and p-values
prerequisite:
dependencies: ctree from library party
}
