% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectSize.R
\name{effectSize}
\alias{effectSize}
\title{get aboslute fitness for a rule}
\usage{
effectSize(subgroup.data, y_idx, y_type = NULL, subgroup_idx)
}
\arguments{
\item{y_idx}{index of outcome to calculate measure on (integer or string)}

\item{subgroup_idx}{a logical vector indicating participation in subgroup}

\item{subgroup.dataset}{a list of data, covariates, outcomes and the trt, see loadDataset (list)}
}
\value{
c(effect_size)
}
\description{
This function allows you to calculate the absolute fitness for a given rule.
Fitness is what we try to minimize/maximize. Fitness in this version is
cohen's d for numeric values and cohen's h for binary.
}
\details{
prerequisite:
}
