% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploreCorrelationMatrix.R
\name{exploreCorrelationMatrix}
\alias{exploreCorrelationMatrix}
\title{explore the correlations between the features}
\usage{
exploreCorrelationMatrix(subgroup.data, method = "pearson",
  significance = 0.025)
}
\arguments{
\item{method}{How to calculate the correlation c("pearson", "spearman")}

\item{subgroup.dataset}{a list of data, covariates, outcomes and the trt, see loadDataset (list)}

\item{significant}{P-value threshold (numeric)}
}
\value{
plots the correlation matrix using corrplot
}
\description{
This function allows you to plot the correlation between features
with either pearson or spearson method
using the corrplot function, which displays correlations and p-values
prerequisite: exploreCorrelationMatrix
dependencies: corrplot
}
