% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploreDetailedChartCorrelation.R
\name{exploreDetailedChartCorrelation}
\alias{exploreDetailedChartCorrelation}
\title{explore the correlations between the features}
\usage{
exploreDetailedChartCorrelation(subgroup.data)
}
\arguments{
\item{subgroup.dataset}{a list of data, covariates, outcomes and the trt, see loadDataset (list)}
}
\value{
plots the correlation matrix using corrplot
}
\description{
This function allows you to plot a detail plot of correlation between features
with PerformanceAnalytics. However all fields will be converted to numeric
prerequisite:
dependencies: PerformanceAnalytics
}
