% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploreTriangleCorrelationMatrix.R
\name{exploreTriangleCorrelationMatrix}
\alias{exploreTriangleCorrelationMatrix}
\title{explore the correlations between the features}
\usage{
exploreTriangleCorrelationMatrix(subgroup.data, method = "pearson")
}
\arguments{
\item{method}{How to calculate the correlation c("pearson", "spearman")}

\item{subgroup.dataset}{a list of data, covariates, outcomes and the trt, see loadDataset (list)}

\item{significant}{P-value threshold (numeric)}
}
\value{
plots the correlation matrix using corrplot
}
\description{
This function allows you to plot the correlation between features
with either pearson or spearson method using ggplot2
prerequisite: exploreCorrelationMatrix
dependencies: reshape, ggplot2
}
