% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseTSDTResults.R
\name{parseTSDTResults}
\alias{parseTSDTResults}
\title{parseTSDTResults
parse TSDT table result return rule structure}
\usage{
parseTSDTResults(TSDT_table, outcome, param, filter = "Strong")
}
\arguments{
\item{TSDT_table}{a data.table of results returned by TSDT [data.table]}

\item{outcome}{name of outcome column that TSDT was run on [character]}

\item{parameters}{used to run TSDT, saved as log [list or character]}
}
\value{
a rule structure [data.table]
}
\description{
This function parses a data.table returned by runTSDT.
A rule structure has at least 5 columns: rule, method, outcome, desc, and param
The rule can be parsed by parseRule to get subgroup
}
