% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseVTResults.R
\name{parseVTResults}
\alias{parseVTResults}
\title{parseVTResults
parse Virtual Twins table result return rule structure}
\usage{
parseVTResults(VT_table, outcome, param)
}
\arguments{
\item{VT_table}{a data.table of results returned by TSDT [data.table]}

\item{outcome}{name of outcome column that TSDT was run on [character]}

\item{parameters}{used to run TSDT, saved as log [list or character]}
}
\value{
a rule structure [data.table]
}
\description{
This function parses a data.table returned by runTSDT.
A rule structure has at least 5 columns: rule, method, outcome, desc, and param
The rule can be parsed by parseRule to get subgroup
}
